requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
jinja2==3.1.2
weasyprint==60.1
pydantic==2.5.2
loguru==0.7.2 